Hide Interact if not enabled
The option to separate the Interact action from Enter is available to
users from the speedrunning options menu, but is not set by default.
This may lead to some confusion for users when they see the Interact
bindings in the controller menu but there is not action taken in-game.
Making it so that the Interact binding option in the menu is hidden if
the separate_interact is not on.